+++ src/openpose/net/maximumCaffe.cpp
@@ -28,8 +28,8 @@ namespace op
     }
 
     template <typename T>
-    void MaximumCaffe<T>::LayerSetUp(const std::vector<caffe::Blob<T>*>& bottom,
-                                     const std::vector<caffe::Blob<T>*>& top)
+    void MaximumCaffe<T>::LayerSetUp(const std::vector<ArrayCpuGpu<T>*>& bottom,
+                                     const std::vector<ArrayCpuGpu<T>*>& top)
     {
         try
         {
@@ -50,8 +50,8 @@ namespace op
     }
 
     template <typename T>
-    void MaximumCaffe<T>::Reshape(const std::vector<caffe::Blob<T>*>& bottom,
-                                  const std::vector<caffe::Blob<T>*>& top)
+    void MaximumCaffe<T>::Reshape(const std::vector<ArrayCpuGpu<T>*>& bottom,
+                                  const std::vector<ArrayCpuGpu<T>*>& top)
     {
         try
         {
@@ -86,8 +86,8 @@ namespace op
     }
 
     template <typename T>
-    void MaximumCaffe<T>::Forward(const std::vector<caffe::Blob<T>*>& bottom,
-                                  const std::vector<caffe::Blob<T>*>& top)
+    void MaximumCaffe<T>::Forward(const std::vector<ArrayCpuGpu<T>*>& bottom,
+                                  const std::vector<ArrayCpuGpu<T>*>& top)
     {
         try
         {
@@ -108,8 +108,8 @@ namespace op
     }
 
     template <typename T>
-    void MaximumCaffe<T>::Forward_cpu(const std::vector<caffe::Blob<T>*>& bottom,
-                                      const std::vector<caffe::Blob<T>*>& top)
+    void MaximumCaffe<T>::Forward_cpu(const std::vector<ArrayCpuGpu<T>*>& bottom,
+                                      const std::vector<ArrayCpuGpu<T>*>& top)
     {
         try
         {
@@ -127,8 +127,8 @@ namespace op
     }
 
     template <typename T>
-    void MaximumCaffe<T>::Forward_gpu(const std::vector<caffe::Blob<T>*>& bottom,
-                                      const std::vector<caffe::Blob<T>*>& top)
+    void MaximumCaffe<T>::Forward_gpu(const std::vector<ArrayCpuGpu<T>*>& bottom,
+                                      const std::vector<ArrayCpuGpu<T>*>& top)
     {
         try
         {
@@ -148,9 +148,9 @@ namespace op
     }
 
     template <typename T>
-    void MaximumCaffe<T>::Backward_cpu(const std::vector<caffe::Blob<T>*>& top,
+    void MaximumCaffe<T>::Backward_cpu(const std::vector<ArrayCpuGpu<T>*>& top,
                                        const std::vector<bool>& propagate_down,
-                                       const std::vector<caffe::Blob<T>*>& bottom)
+                                       const std::vector<ArrayCpuGpu<T>*>& bottom)
     {
         try
         {
@@ -168,9 +168,9 @@ namespace op
     }
 
     template <typename T>
-    void MaximumCaffe<T>::Backward_gpu(const std::vector<caffe::Blob<T>*>& top,
+    void MaximumCaffe<T>::Backward_gpu(const std::vector<ArrayCpuGpu<T>*>& top,
                                        const std::vector<bool>& propagate_down,
-                                       const std::vector<caffe::Blob<T>*>& bottom)
+                                       const std::vector<ArrayCpuGpu<T>*>& bottom)
     {
         try
         {